// src/app/portal/moduls/my-account/my-invoices/models/list-invoices.ts
export class ListInvoices {
    Amount: string;
    BalanceInvoice: string;
    BroadcastDate: string;
    Currency: string;
    DescriptionInvoice: string;
    ExpirationDate: string;
    Finalconsumer: string;
    InternalNumber: string;
    InternalSeries: string;
    Number: string;
    Series: string;
    State: string;
    Taxedamount: string;
    VoucherType: string;
}